// generated/models/Group.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Group
 */
export interface Group {
    /**
     * 
     * @type {string}
     * @memberof Group
     */
    groupNumber: string;
    /**
     * 
     * @type {number}
     * @memberof Group
     */
    courseNumber: number;
    /**
     * 
     * @type {number}
     * @memberof Group
     */
    id: number;
}

/**
 * Check if a given object implements the Group interface.
 */
export function instanceOfGroup(value: object): value is Group {
    if (!('groupNumber' in value) || value['groupNumber'] === undefined) return false;
    if (!('courseNumber' in value) || value['courseNumber'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function GroupFromJSON(json: any): Group {
    return GroupFromJSONTyped(json, false);
}

export function GroupFromJSONTyped(json: any, ignoreDiscriminator: boolean): Group {
    if (json == null) {
        return json;
    }
    return {
        
        'groupNumber': json['groupNumber'],
        'courseNumber': json['courseNumber'],
        'id': json['id'],
    };
}

export function GroupToJSON(value?: Group | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'groupNumber': value['groupNumber'],
        'courseNumber': value['courseNumber'],
        'id': value['id'],
    };
}