// generated/models/StudentWithGroup.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Group } from './Group';
import {
    GroupFromJSON,
    GroupFromJSONTyped,
    GroupToJSON,
} from './Group';

/**
 * 
 * @export
 * @interface StudentWithGroup
 */
export interface StudentWithGroup {
    /**
     * 
     * @type {string}
     * @memberof StudentWithGroup
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof StudentWithGroup
     */
    groupId: number;
    /**
     * 
     * @type {boolean}
     * @memberof StudentWithGroup
     */
    status: boolean;
    /**
     * 
     * @type {number}
     * @memberof StudentWithGroup
     */
    id: number;
    /**
     * 
     * @type {Date}
     * @memberof StudentWithGroup
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof StudentWithGroup
     */
    updatedAt: Date;
    /**
     * 
     * @type {Group}
     * @memberof StudentWithGroup
     */
    group: Group;
}

/**
 * Check if a given object implements the StudentWithGroup interface.
 */
export function instanceOfStudentWithGroup(value: object): value is StudentWithGroup {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('groupId' in value) || value['groupId'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    if (!('group' in value) || value['group'] === undefined) return false;
    return true;
}

export function StudentWithGroupFromJSON(json: any): StudentWithGroup {
    return StudentWithGroupFromJSONTyped(json, false);
}

export function StudentWithGroupFromJSONTyped(json: any, ignoreDiscriminator: boolean): StudentWithGroup {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'groupId': json['groupId'],
        'status': json['status'],
        'id': json['id'],
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
        'group': GroupFromJSON(json['group']),
    };
}

export function StudentWithGroupToJSON(value?: StudentWithGroup | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'groupId': value['groupId'],
        'status': value['status'],
        'id': value['id'],
        'createdAt': ((value['createdAt']).toISOString()),
        'updatedAt': ((value['updatedAt']).toISOString()),
        'group': GroupToJSON(value['group']),
    };
}